- Added status property to OMembership managed object class, making redundant   the earler isActive flag, which has been removed. - Working on getting native iOS localisation to work, this commit mainly to   sync up git after upgrading to Xcode 5.1.

+++ OrigoApp/OConnection.m
@@ -309,7 +309,7 @@ static NSString * const kURLParameterIdentifier = @"id";
 
 - (void)connectionDidFinishLoading:(NSURLConnection *)connection;
 {
-    OLogDebug(@"Server request completed. HTTP status code: %d", _HTTPResponse.statusCode);
+    OLogDebug(@"Server request completed. HTTP status code: %ld", _HTTPResponse.statusCode);
     
     [UIApplication sharedApplication].networkActivityIndicatorVisible = NO;
 
@@ -339,7 +339,7 @@ static NSString * const kURLParameterIdentifier = @"id";
 
 - (void)connection:(NSURLConnection *)connection didFailWithError:(NSError *)error
 {
-	OLogError(@"Connection failed with error: %@ (%d)", error.localizedDescription, error.code);
+	OLogError(@"Connection failed with error: %@ (%ld)", error.localizedDescription, (long)error.code);
     
     [UIApplication sharedApplication].networkActivityIndicatorVisible = NO;
